api_tools/requirements.txt
Django==1.11.3
djangorestframework==3.6.3
djangorestframework-jsonapi==2.2.0
djangorestframework-jwt==1.11.0
django-debug-toolbar==1.8
django-environ==0.4.3
django-extensions==1.8.0
django-jet==1.0.6
draceditor==1.1.8
